shared/requirements.txt
azure-identity
azure-keyvault-secrets
python-dotenv
requests
